Add lendings synchronization

+++ composeApp/src/phonesMain/kotlin/org/centrexcursionistalcoi/app/database/DepartmentsRepository.phonesMain.kt
@@ -2,4 +2,4 @@ package org.centrexcursionistalcoi.app.database
 
 import org.centrexcursionistalcoi.app.data.Department
 
-actual val DepartmentsRepository: Repository<Department, Long> = DepartmentsDatabaseRepository
+actual val DepartmentsRepository: Repository<Department, Int> = DepartmentsDatabaseRepository
